requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
pydantic==2.9.2
web3==7.2.0
python-dotenv==1.0.1